requirements.txt
Flask~=2.0.2
pytest
python-dotenv~=0.19.2
flask_sqlalchemy
SQLAlchemy~=1.4.31